Let indices in sse_trans be ssize_t.

+++ bitmath.c
@@ -19,7 +19,7 @@ sse_trans(uint8_t const *inp, uint8_t *out, int nrows, int ncols)
 {
 #   define INP(x,y) inp[(x)*ncols/8 + (y)/8]
 #   define OUT(x,y) out[(y)*nrows/8 + (x)/8]
-  int rr, cc, i, h;
+  ssize_t rr, cc, i, h;
   union { __m128i x; uint8_t b[16]; } tmp;
   assert(nrows % 8 == 0 && ncols % 8 == 0);
